+++ apps/react-app/src/App.tsx
@@ -1,15 +1,27 @@
-import { useResizeObserver } from "@ianduvall/react-resize-observer-store";
+import {
+	useResizeObserver,
+	type Entry,
+} from "@ianduvall/react-resize-observer-store";
+import { useCallback } from "react";
 import "./App.css";
 
-const selector = (entry: ResizeObserverEntry | undefined) => {
-	if (!entry) return undefined;
-	const { blockSize, inlineSize } = entry.contentBoxSize[0];
-	return { blockSize, inlineSize };
-};
-
 function App() {
+	const selector = useCallback((entry: Entry | undefined) => {
+		if (!entry) return undefined;
+		const boxSize = entry.borderBoxSize[0];
+		return {
+			blockSize: boxSize.blockSize,
+			inlineSize: boxSize.inlineSize,
+		};
+	}, []);
+	const isEqual = useCallback((a: ReturnType<typeof selector>, b: typeof a) => {
+		if (!a || !b) return false;
+		return a.blockSize === b.blockSize && a.inlineSize === b.inlineSize;
+	}, []);
 	const [ref, selection] = useResizeObserver({
+		box: "border-box",
 		selector,
+		isEqual,
 	});
 	console.log(selection);