+++ packages/lazily/tsup.config.js
@@ -3,6 +3,7 @@ import { defineConfig } from "tsup";
 export default defineConfig({
 	name: "resize-observer-store",
 
+	clean: true,
 	dts: true,
 	entry: ["src/index.ts"],
 	format: ["cjs", "esm"],