+++ packages/react-resize-observer-store/src/use-resize-observer.ts
@@ -1,24 +1,32 @@
-import { getResizeObserverStore } from "@ianduvall/resize-observer-store";
 import { useCallback, useMemo, useState } from "react";
 import { useSyncExternalStoreWithSelector } from "use-sync-external-store/with-selector";
+import {
+	getResizeObserverStore,
+	type Entry,
+	type BoxOption,
+} from "@ianduvall/resize-observer-store";
 
-type BoxOption = "border-box" | "content-box" | "device-pixel-content-box";
+export type { Entry, BoxOption } from "@ianduvall/resize-observer-store";
 
-const identity = <Input>(input: Input): Input => input;
+export type Selector<Elem extends Element, Selection> = (
+	entry: Entry<Elem>,
+) => Selection;
+export type IsEqual<Selection> = (a: Selection, b: Selection) => boolean;
 
-export const useResizeObserver = <Selection = ResizeObserverEntry | undefined>({
+export const useResizeObserver = <
+	Elem extends Element,
+	Selection = Entry<Elem>,
+>({
 	box = "content-box" as const,
-	selector = identity as (entry: ResizeObserverEntry | undefined) => Selection,
+	selector,
 	isEqual,
 }: {
 	box?: BoxOption;
-	selector?: (entry: ResizeObserverEntry | undefined) => Selection;
-	isEqual?: (a: Selection, b: Selection) => boolean;
-} = {}) => {
-	const [element, setElement] = useState<Element | null>(null);
-	const refCallback = useCallback((node: Element | null) => {
-		setElement(node);
-	}, []);
+	selector?: Selector<Elem, Selection>;
+	isEqual?: IsEqual<Selection>;
+} = {}): readonly [(el: Elem | null) => void, Selection | undefined] => {
+	const [element, setElement] = useState<Elem | null>(null);
+	const refCallback = setElement;
 
 	const subscribe = useCallback(
 		(onStoreChanged: () => void) => {
@@ -49,10 +57,38 @@ export const useResizeObserver = <Selection = ResizeObserverEntry | undefined>({
 		return store.getSnapshot(element, box);
 	}, [box, element]);
 
+	/**
+	 * Wrap selector to prevent it from being called with undefined.
+	 */
+	const selectorWrapper = useCallback(
+		(entry: Entry<Elem> | undefined) => {
+			if (!entry) {
+				return;
+			}
+			return selector?.(entry);
+		},
+		[selector],
+	);
+
+	/**
+	 * Wrap isEqual to prevent it from being called with undefined.
+	 */
+	const isEqualWrapper = useMemo(() => {
+		if (!isEqual) {
+			return;
+		}
+		return (a: Selection | undefined, b: Selection | undefined) => {
+			if (a === undefined || b === undefined) {
+				return a === b;
+			}
+			return isEqual(a, b);
+		};
+	}, [isEqual]);
+
 	const selection = useSyncExternalStoreWithSelector<
-		ResizeObserverEntry | undefined,
-		Selection
-	>(subscribe, getSnapshot, undefined, selector, isEqual);
+		Entry<Elem> | undefined,
+		Selection | undefined
+	>(subscribe, getSnapshot, undefined, selectorWrapper, isEqualWrapper);
 
 	return useMemo(
 		() => [refCallback, selection] as const,