+++ packages/resize-observer-store/src/index.ts
@@ -1,6 +1,8 @@
 import { ResizeObserverStore } from "./resize-observer-store";
 import { lazily } from "@ianduvall/lazily";
 
+export * from "./types";
+
 /**
  * Returns the singleton instance of the ResizeObserverStore.
  * @returns ResizeObserverStore